Backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.7
psycopg3==3.1.12
pgvector==0.2.3

# Security
bcrypt==4.0.1
PyJWT==2.8.0

# AI/ML
openai==1.3.0
sentence-transformers==2.2.2
nltk==3.8.1
rapidfuzz==3.4.0
numpy==1.24.3
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
celery==5.3.1

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Documentation
flasgger==0.9.7.1
marshmallow==3.20.1

# Monitoring and logging
structlog==23.1.0
prometheus-client==0.17.1
